refactor(assets): refactor assets imports/exports
