+++ assets/index.js
@@ -1,0 +1,12 @@
+export { default as IcComponent } from './component.svg'
+export { default as DownArrow } from './down-arrow.svg'
+export { default as GithubLogo } from './github-logo.svg'
+export { default as Google } from './google.svg'
+export { default as IcGuide } from './guide.svg'
+export { default as PurpleShape } from './purple-shape.svg'
+export { default as IcResources } from './resources.svg'
+export { default as SampleLogo } from './sample-logo.svg'
+export { default as SearchIcon } from './search.svg'
+export { default as SpacePeople } from './space-people.svg'
+export { default as Stars } from './stars.svg'
+export { default as TipeLogo } from './tipe-logo.svg'
