+++ pages/resource.vue
@@ -1,0 +1,19 @@
+<template>
+  <component-grid>
+    <h1>Resource</h1>
+  </component-grid>
+</template>
+
+<script>
+import ComponentGrid from '~/components/Component/Grid.vue'
+
+export default {
+  layout: 'docs',
+  components: {
+    ComponentGrid
+  }
+}
+</script>
+
+<style lang="postcss" scoped>
+</style>
